photos_api/requirements.txt
Flask==2.0.2
Pillow==8.4.0
Flask-Cors==3.0.10
psycopg2-binary==2.9.1